DOC-7979: tidy Error Messages Reference
The error code numbers were taken from e.g.

https://github.com/couchbaselabs/sdk-rfcs/blob/master/rfc/0058-error-handling.md

but have changed in the source libcouchbase:
for example "DocumentExists" was originally 105, but has now become 305.
This is because the number itself is simply an implementation detail of
the libcouchbase library (and isn't exposed in Java SDK for example.)

The name of the error is the key thing to focus on.

This commit then:

- removes the old (wrong) libcouchbase error numbers.
- reformats for consistency
- fixes some Asciidoc errors (lists that weren't separated by a blank line) etc.